Unnecessary casts
git-svn-id: https://svn.apache.org/repos/asf/commons/proper/lang/trunk@753656 13f79535-47bb-0310-9956-ffa450edef68

+++ src/test/org/apache/commons/lang/math/RandomUtilsTest.java
@@ -331,8 +331,8 @@ public final class RandomUtilsTest extends TestCase {
         double sumSq = 0.0d;
         double dev = 0.0d;
         for (int i = 0; i< observed.length; i++) {
-            dev = (double)(observed[i] - expected[i]);
-            sumSq += dev*dev/(double)expected[i];
+            dev = (observed[i] - expected[i]);
+            sumSq += dev*dev/expected[i];
         }
         return sumSq;
     }
